requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.31
pydantic==2.4.2
python-dotenv==1.0.0
